Answer Check for Updates in the Manager, rather than asking for a machine
Help > Check for Updates said "Start a machine first - this command acts
on the machine being shown". It does not: it compares this build's version
against the newest release and offers the release page. With no machine
running there was no way to reach it at all.

The Manager answers it itself, as it already does for About and the
website links - and has to, a managed machine having no window for the
progress dialogue and message boxes to appear over.

So the check moves out of MainFrame into check_update.cpp, with the
version comparison and the release-JSON reader it uses, and both windows
call it. The same arrangement as AboutDialog, for the same reason: a
second copy written for the Manager would be wrong the first time either
changed.

The body is unchanged - byte-identical to what it replaces apart from
taking its parent window as an argument.

Create Support Bundle stays behind the guard. That one really does act on
the machine being shown: it screenshots its panel and collects its log.
